+++ tests/run_all.py
@@ -30,6 +30,7 @@ if __name__ == '__main__':
         loader.discover(HERE, pattern='test_audio.py'),
         loader.discover(HERE, pattern='test_update.py'),
         loader.discover(HERE, pattern='test_sessions.py'),
+        loader.discover(HERE, pattern='test_wrap.py'),
         loader.discover(HERE, pattern='test_settings.py'),
         loader.discover(HERE, pattern='test_e2e.py'),
         loader.discover(HERE, pattern='test_watch.py'),